backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pydantic
pydantic[email]
email-validator
python-multipart
sqlalchemy
bcrypt
python-jose[cryptography]
